misc: move routes registration to its own file
